init: move record_once into finalize_app
