+++ invenio_communities/views/api.py
@@ -17,31 +17,6 @@ blueprint = Blueprint(
 )
 
 
-@blueprint.record_once
-def init(state):
-    """Init app."""
-    app = state.app
-    # Register services - cannot be done in extension because
-    # Invenio-Records-Resources might not have been initialized.
-    rr_ext = app.extensions["invenio-records-resources"]
-    idx_ext = app.extensions["invenio-indexer"]
-    ext = app.extensions["invenio-communities"]
-
-    # services
-    rr_ext.registry.register(ext.service, service_id="communities")
-    rr_ext.registry.register(ext.service.members, service_id="members")
-
-    # indexers
-    idx_ext.registry.register(ext.service.indexer, indexer_id="communities")
-    idx_ext.registry.register(ext.service.members.indexer, indexer_id="members")
-    idx_ext.registry.register(
-        ext.service.members.archive_indexer, indexer_id="archived-invitations"
-    )
-
-    # change notification handlers
-    rr_ext.notification_registry.register("users", ext.service.on_relation_update)
-
-
 def create_communities_api_blueprint(app):
     """Create communities api blueprint."""
     ext = app.extensions["invenio-communities"]
