+++ setup.cfg
@@ -90,7 +90,7 @@ invenio_users_resources.moderation.actions =
     restore = invenio_communities.requests.user_moderation.actions:on_restore
     approve = invenio_communities.requests.user_moderation.actions:on_approve
 invenio_base.finalize_app =
-    invenio_communities = invenio_communities.finalize_app:finalize_app
+    invenio_communities = invenio_communities.ext:finalize_app
 
 
 [build_sphinx]
